Improved Shutdown Hooks

+++ src/test/java/ca/empire/setup/Hooks.java
@@ -1,6 +1,7 @@
 package ca.empire.setup;
 
 import ca.empire.setup.configuration.Config;
+import ca.empire.setup.configuration.models.DownloadManagement;
 import ca.empire.setup.configuration.models.Environment;
 import ca.empire.util.FileOperations;
 import ca.empire.util.TestEnvironment;
@@ -8,9 +9,12 @@ import io.cucumber.java.After;
 import io.cucumber.java.Before;
 import io.cucumber.java.Scenario;
 import io.cucumber.java.Status;
+import org.apache.commons.io.FileUtils;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.openqa.selenium.JavascriptExecutor;
+import org.openqa.selenium.OutputType;
+import org.openqa.selenium.TakesScreenshot;
 import org.openqa.selenium.WebDriver;
 
 import java.io.File;
@@ -51,16 +55,23 @@ public class Hooks {
             throw new RuntimeException(e.getCause());
         }
 
-        Runtime.getRuntime()
-                .addShutdownHook(
-                        new Thread(
-                                () -> {
-                                    logger.traceEntry();
-                                    FileOperations.deleteDir(
-                                            new File(System.getProperty("user.dir") + "/.temp"));
-                                    logger.traceExit();
-                                    LogManager.shutdown(true);
-                                }));
+        DownloadManagement.DeletionCondition deletionCondition =
+                config.getConfigurationModel().downloadManagement.deletionCondition;
+
+        if (deletionCondition == DownloadManagement.DeletionCondition.afterAll
+                || deletionCondition == DownloadManagement.DeletionCondition.afterEach) {
+            Runtime.getRuntime()
+                    .addShutdownHook(
+                            new Thread(
+                                    () -> {
+                                        logger.traceEntry();
+                                        FileOperations.deleteDir(
+                                                new File(
+                                                        System.getProperty("user.dir") + "/.temp"));
+                                        logger.traceExit();
+                                        LogManager.shutdown(true);
+                                    }));
+        }
 
         logger.traceExit();
     }
@@ -162,14 +173,55 @@ public class Hooks {
     /* Standard Hooks */
 
     /**
-     * Takes a screenshot if the scenario failed.
+     * Takes a screenshot if the scenario did not pass.
      *
      * @param scenario - The test scenario
      */
     @After(order = 4)
     public void screenCapture(Scenario scenario) {
         logger.traceEntry();
-        /* ... */
+
+        if (getDriver() == null) {
+            logger.traceExit("Driver is null");
+            return;
+        } else if (scenario.getStatus() == Status.PASSED) {
+            logger.traceExit("Scenario has passed");
+            return;
+        }
+
+        TakesScreenshot screenshotDriver;
+
+        try {
+            screenshotDriver = (TakesScreenshot) getDriver();
+        } catch (ClassCastException e) {
+            logger.traceExit("Driver is unable to take screenshots");
+            return;
+        }
+
+        String filename = scenario.getId() + ".png";
+        File screenshotFile = screenshotDriver.getScreenshotAs(OutputType.FILE);
+
+        // Save screenshot locally
+        try {
+            File savedFile =
+                    new File("./build/testResults/" + scenario.getStatus() + "/" + filename);
+            FileUtils.copyFile(screenshotFile, savedFile);
+        } catch (IOException e) {
+            logger.warn(
+                    "An exception occurred while trying to save the screenshot: {}",
+                    e.getMessage());
+        }
+
+        // Attach screenshot to the scenario
+        try {
+            byte[] screenshotBytes = Files.readAllBytes(screenshotFile.toPath());
+            scenario.attach(screenshotBytes, "image/png", filename);
+        } catch (IOException e) {
+            logger.warn(
+                    "An exception occurred while trying to attach the screenshot: {}",
+                    e.getMessage());
+        }
+
         logger.traceExit();
     }
 
@@ -233,6 +285,7 @@ public class Hooks {
     public void deleteDownloadDirectory() {
         logger.traceEntry();
         File downloadDir = getDownloadDirectory();
+        DownloadManagement downloadManagement = config.getConfigurationModel().downloadManagement;
 
         if (downloadDir == null) {
             logger.info("This driver does not have any download directories associated with it.");
@@ -245,35 +298,46 @@ public class Hooks {
         long maxSize = (long) Math.pow(2, 23); // roughly 8MB
 
         if (downloadedFiles == null) {
-            logger.warn("Attempting to retrieve the list of downloaded files returned a null.");
+            logger.warn("Attempting to retrieve the list of downloaded files returned a null value.");
             logger.traceExit();
             return;
         }
 
-        for (File downloadedFile : downloadedFiles) {
-            long size = downloadedFile.length();
+        DownloadManagement.AttachmentCondition condition = downloadManagement.attachmentCondition;
+        boolean attachFiles =
+                (condition == DownloadManagement.AttachmentCondition.failure && scenario.isFailed())
+                        || (condition == DownloadManagement.AttachmentCondition.unsuccessful
+                                && scenario.getStatus() != Status.PASSED)
+                        || condition == DownloadManagement.AttachmentCondition.always;
 
-            if (size > maxSize) {
-                double convertedSize = size / Math.pow(2, 20); // convert to MB
-                logger.warn(
-                        downloadedFile.getAbsolutePath()
-                                + " exceeds max size ("
-                                + convertedSize
-                                + "MB)");
-            } else {
-                try {
-                    byte[] fileBytes = Files.readAllBytes(downloadedFile.toPath());
-                    String mimeType = Files.probeContentType(downloadedFile.toPath());
+        if (attachFiles) {
+            for (File downloadedFile : downloadedFiles) {
+                long size = downloadedFile.length();
 
-                    scenario.attach(
-                            fileBytes, mimeType, downloadedFile.getName());
-                } catch (Exception e) {
-                    logger.warn(e);
+                if (size > maxSize) {
+                    double convertedSize = size / Math.pow(2, 20); // convert to MB
+                    logger.warn(
+                            downloadedFile.getAbsolutePath()
+                                    + " exceeds max size ("
+                                    + convertedSize
+                                    + "MB)");
+                } else {
+                    try {
+                        byte[] fileBytes = Files.readAllBytes(downloadedFile.toPath());
+                        String mimeType = Files.probeContentType(downloadedFile.toPath());
+
+                        scenario.attach(fileBytes, mimeType, downloadedFile.getName());
+                    } catch (Exception e) {
+                        logger.warn(e);
+                    }
                 }
             }
         }
 
-        FileOperations.deleteDir(downloadDir);
+        if (downloadManagement.deletionCondition
+                == DownloadManagement.DeletionCondition.afterEach) {
+            FileOperations.deleteDir(downloadDir);
+        }
         logger.traceExit();
     }
 
@@ -304,6 +368,13 @@ public class Hooks {
         testEnvironments.remove();
 
         if (driverDecorators.get() != null) {
+            try {
+                getDriver().close();
+            } catch (Exception e) {
+                logger.warn(
+                        "An exception occurred while trying to close the driver: {}",
+                        e.getMessage());
+            }
             getDriver().quit();
             driverDecorators.remove();
         }
@@ -319,6 +390,9 @@ public class Hooks {
     ====================================================================================================================
      */
 
+    /**
+     * @return the download directory assigned to this test runner.
+     */
     public static File getDownloadDirectory() {
         logger.traceEntry();
         File dir = driverDecorators.get().getDownloadDirectory();
@@ -326,6 +400,9 @@ public class Hooks {
         return dir;
     }
 
+    /**
+     * @return the driver assigned to this test runner.
+     */
     public static WebDriver getDriver() {
         logger.traceEntry();
         WebDriver driver = driverDecorators.get().getDriver();
@@ -333,24 +410,38 @@ public class Hooks {
         return driver;
     }
 
+    /**
+     * @return the test environment assigned to this test runner.
+     */
     public static TestEnvironment getTestEnvironment() {
         logger.traceEntry();
         logger.traceExit(testEnvironments.get());
         return testEnvironments.get();
     }
 
+    /**
+     * @return the scenario assigned to this test runner.
+     */
     public static Scenario getScenario() {
         logger.traceEntry();
         logger.traceExit(scenarios.get());
         return scenarios.get();
     }
 
+    /**
+     * @return the parsed configuration that is used by this program.
+     */
     public Config getConfig() {
         logger.traceEntry();
         logger.traceExit(config);
         return config;
     }
 
+    /**
+     * Sets the DriverDecorator that this test runner will use.
+     *
+     * @param driver The driver decorator that this test runner will use.
+     */
     private static void setDriverDecorator(DriverDecorator driver) {
         logger.traceEntry(() -> driver);
         driverDecorators.set(driver);
